requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
sqlalchemy==1.4.22
pymysql
pytz
pandas
mysql
mysql-connector
openpyxl